Revert "Merge branch 'admission_v3' of https://github.com/APC-SoCIT/APC-2024-2025-T1-05-Admission-Management-System into admission_v3"
This reverts commit de305f8e6762c6f5548bfda295ed47cb5cf00e2c, reversing
changes made to 435f688849c403dbf39cf9449c43a349c044bce5.

+++ InnolabAMS/database/migrations/2025_01_09_155018_add_status_to_applicant_infos.php
@@ -17,8 +17,11 @@ return new class extends Migration
 
     public function down()
     {
-        Schema::table('applicant_infos', function (Blueprint $table) {
-            $table->dropColumn('status');
-        });
+        if (Schema::hasColumn('applicant_infos', 'status')) {
+            Schema::table('applicant_infos', function (Blueprint $table) {
+                $table->dropColumn('status');
+            });
+        }
     }
+    
 };
